vertical slides draft
This is a custom-ish slide subsystem class. It uses PID, power caching, and has manual control for easily tuning set poses. It is far from perfect, so don't look too deep or try to copy exactly. This is just an example of how one could implement custom PID and power caching